+++ tests/src/rules/no-anonymous-default-export.js
@@ -22,6 +22,7 @@ ruleTester.run('no-anonymous-default-export', rule, {
     test({ code: 'export default `foo`', options: [{ allowLiteral: true }] }),
     test({ code: 'export default {}', options: [{ allowObject: true }] }),
     test({ code: 'export default foo(bar)', options: [{ allowCallExpression: true }] }),
+    test({ code: 'export default new Foo()', options: [{ allowNew: true }] }),
 
     // Allow forbidden types with multiple options
     test({ code: 'export default 123', options: [{ allowLiteral: true, allowObject: true }] }),
@@ -53,6 +54,7 @@ ruleTester.run('no-anonymous-default-export', rule, {
     test({ code: 'export default `foo`', errors: [{ message: 'Assign literal to a variable before exporting as module default' }] }),
     test({ code: 'export default {}', errors: [{ message: 'Assign object to a variable before exporting as module default' }] }),
     test({ code: 'export default foo(bar)', options: [{ allowCallExpression: false }], errors: [{ message: 'Assign call result to a variable before exporting as module default' }] }),
+    test({ code: 'export default new Foo()', errors: [{ message: 'Assign instance to a variable before exporting as module default' }] }),
 
     // Test failure with non-covering exception
     test({ code: 'export default 123', options: [{ allowObject: true }], errors: [{ message: 'Assign literal to a variable before exporting as module default' }] }),